requirements.txt
mss>=9.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
